fix: ldat存在的一些问题

+++ include_private/common/DAC.hpp
@@ -7,13 +7,13 @@
 namespace Union::Base {
     class DAC {
     public:
-        double               baseGain         = {}; ///< 基础增益
-        double               biasGain         = {}; ///< 补偿增益
-        int                  gate             = {};
-        bool                 isSubline        = {};
-        std::vector<int>     index            = {}; ///< 索引
-        std::vector<uint8_t> value            = {}; ///< 值
-        int                  samplingDepth    = {}; ///< 采样深度
-        double               decimationFactor = {}; ///< 采样因子
+        double              baseGain         = {}; ///< 基础增益
+        double              biasGain         = {}; ///< 补偿增益
+        int                 gate             = {};
+        bool                isSubline        = {};
+        std::vector<double> index            = {}; ///< 索引
+        std::vector<double> value            = {}; ///< 值
+        int                 samplingDepth    = {}; ///< 采样深度
+        double              decimationFactor = {}; ///< 采样因子
     };
 } // namespace Union::Base
